+++ SandboxiePlus/SandMan/Helpers/WinHelper.cpp
@@ -10,14 +10,14 @@
 #include <Shlwapi.h>
 #include <Shlobj.h>
 
-QImage LoadWindowsIcon(const QString& Path, quint32 Index)
+QPixmap LoadWindowsIcon(const QString& Path, quint32 Index)
 {
 	std::wstring path = QString(Path).replace("/", "\\").toStdWString();
 	HICON icon = ExtractIconW(NULL, path.c_str(), Index);
 #if QT_VERSION < QT_VERSION_CHECK(6, 0, 0)
-	QImage Icon = QtWin::fromHICON(icon).toImage();
+	QPixmap Icon = QtWin::fromHICON(icon);
 #else
-	QImage Icon = QImage::fromHICON(icon);
+	QPixmap Icon = QPixmap::fromImage(QImage::fromHICON(icon));
 #endif
 	DestroyIcon(icon);
 	return Icon;