+++ SandboxiePlus/SandMan/Helpers/WinHelper.h
@@ -1,5 +1,5 @@
 #pragma once
 
-QImage LoadWindowsIcon(const QString& Path, quint32 Index);
+QPixmap LoadWindowsIcon(const QString& Path, quint32 Index);
 
 bool PickWindowsIcon(QWidget* pParent, QString& Path, quint32& Index);